Made possible to let the dog executable load Lua DLL from its resource
This is only supported when building dogfood for Windows.
When statically linked with Lua support for C modules is effectively disabled. C modules will likely require a Lua DLL that conflicts with the statically linked Lua instance in dogfood. In this case a 'multiple Lua VMs detected' message will be displayed in the console output when you run the resulting executable.

I really liked the single executable approach for Windows to simplify the distribution of dogfood and its derived executables.
The C module issue is fixed by dynamically link to Lua and embed the Lua DLL in the executable's resource load it at initialization.